minor changes to working version

+++ Compose.cpp
@@ -296,7 +296,7 @@ int keypair(unsigned char *sk, unsigned long *sklen,
 	return (0);
 }
 
-int shortmessagesigned(unsigned char *m, unsigned long *mlen,
+int shortmessagesigned(unsigned char *m, unsigned long long *mlen,
 					   const unsigned char *sm, unsigned long smlen,
 					   const unsigned char *pk, unsigned long pklen)
 {
@@ -336,7 +336,7 @@ int shortmessagesigned(unsigned char *m, unsigned long *mlen,
 	return (eof);
 }
 
-int signedshortmessage(unsigned char *sm, unsigned long *smlen,
+int signedshortmessage(unsigned char *sm, unsigned long long *smlen,
 					   const unsigned char *m, unsigned long mlen,
 					   const unsigned char *sk, unsigned long sklen)
